requirements.txt.txt
flask
torch
torchvision
pillow
scikit-learn
matplotlib
numpy
werkzeug